Added some stuff

+++ src/engine/game/game.c
@@ -18,8 +18,14 @@
 #include "utility/utility.h"
 
 static Element *status_line;
+static int is_gameover;
+
+void terminate() {
+    is_gameover = 1;
+}
 
 int initGame() {
+    is_gameover = 0;
     frame_count = 0;
 
     initLog();
@@ -31,26 +37,30 @@ int initGame() {
     initScreenManager();
     writeLog(LOG_INIT, "game::initGame(): Initialized game");
 
-    signal(SIGINT, closeGame);
+    signal(SIGINT, terminate);
+    // signal(SIGSEGV, closeGame);
 
     status_line = NULL;
 
     return 0;
 }
 
-void closeGame(int a) {
+void closeGame() {
     closeScreenManager();
     closeInput();
     closeUi();
     closeObjects();
     closeLog();
 
-    exit(a);
+    exit(0);
 }
 
 void useStatusLine() {
     status_line = createElement();
     status_line->bg_c = COLOR_BLACK;
+    status_line->bg_c_focus = COLOR_RED;
+    status_line->focusable = 1;
+
     registerUiElement(status_line);
 }
 
@@ -63,10 +73,10 @@ int loop() {
     int loop_time = 0;
     int loop_time_saved = 0;
     frame_count = 0;
+    Screen *screen = &screen_manager.main_screen;
+    screen->times_init = 0;
 
-    while (1) {
-        fflush(stdout);
-        // !this->game_over) {
+    while (!is_gameover) {
         // Reset clock
         delta(&game_clock);
         frame_count++;
@@ -79,61 +89,68 @@ int loop() {
         // Retrieve input
         // df::InputManager &input_manager = df::InputManager::getInstance();
         // input_manager.getInput();
-        readInput();
         writeLog(LOG_GAME_V, "game::loop(): Reading input");
+        readInput();
+
+        // Clear the current screen buffer
+        // clearScreen(&screen_manager.main_screen);
+        int i;
 
         // df::WorldManager &world_manager = df::WorldManager::getInstance();
         // Call worldManager update
         // world_manager.update();
+        writeLog(LOG_GAME_V, "game::loop(): Updating objects");
         updateObjects();
-        writeLog(LOG_GAME_V, "game::loop(): Updated objects");
 
         // Send EVENT_BEFOREDRAW to all objects
         // df::EventBeforeDraw p_bd_event = df::EventBeforeDraw();
         // onEvent(&p_bd_event);
+ 
+
+        int elementsRendered;
+        int lightsRendered;
+        int objectsRendered;
 
         // Check that it's okay to write to the buffer
-        sem_wait(&screen_manager.writes_allowed);
+        // Otherwise, skip this frame
+        if (sem_trywait(&screen_manager.writes_allowed)) {
+            for (i = 0; i < screen->ts.lines * screen->ts.cols; i++) {
+                screen->light_buffer[i] = COLOR_BLANK;
+                screen->pixel_buffer[i] = PIXEL_BLANK;
+            }
 
-        // Ui layer is always on top and opaque. 
-        // Ui will render INT_MAX to the depth buffer.
-        int elementsRendered = renderUi();
-        writeLog(LOG_GAME_V, "game::loop(): Rendered %d UI elements", elementsRendered);
+            // Render lighting before rendering objects
+            lightsRendered = renderObjectLights();
+            writeLog(LOG_GAME_V, "game::loop(): Rendered %d lights", lightsRendered);
 
-        // Render lighting before rendering objects
-        int lightsRendered = renderObjectLights();
-        writeLog(LOG_GAME_V, "game::loop(): Rendered %d lights", lightsRendered);
+            objectsRendered = renderObjects();
+            writeLog(LOG_GAME_V, "game::loop(): Rendered %d objects", objectsRendered);
 
-        int objectsRendered = renderObjects();
-        writeLog(LOG_GAME_V, "game::loop(): Rendered %d objects", objectsRendered);
+            // Ui layer is always on top.
+            // Ui will render INT_MAX to the depth buffer.
+            elementsRendered = renderUi();
+            writeLog(LOG_GAME_V, "game::loop(): Rendered %d UI elements", elementsRendered);
 
-        // Let the renderer know that we're done preparing the frame
-        sem_post(&screen_manager.reads_allowed);
+            // Let the renderer know that we're done preparing the frame
+            sem_post(&screen_manager.reads_allowed);
+            writeLog(LOG_GAME_V, "game::loop(): Posted buffer to renderer.");
 
-        // swapScreen();
+            if (status_line != NULL) {
+                String *status_label = createString();
+                sputf(status_label, "Loop: %5d Frame: %5d Obj rendered: %3d Times Init %3d", 
+                        loop_time_saved, 
+                        frame_count, 
+                        objectsRendered,
+                        screen->times_init);
+                writeLog(LOG_GAME_V, "game::loop(): Status %s.", status_label->s);
+                stringCopy(status_line->text, status_label);
+                deleteString(status_label);
+            }
+        }
 
         if (frame_count % 30 == 0) {
             loop_time_saved = loop_time;
         }
-
-        if (status_line != NULL) {
-            String *status_label = createString();
-        // printf("\e[1G\e[%dC", 0);
-            sputf(status_label, " Loop: %5d Frame: %5d Obj rendered: %3d", loop_time_saved, frame_count, objectsRendered);
-        // printf("Frame: %3d ", frame_count);
-        // printf("Inits: %3d ", screen.times_init);
-        // printf("Lines: %3d ", screen.ts.lines);
-        // printf("Cols: %3d ", screen.ts.cols);
-        // printf("Obj rendered: %3d ", objectsRendered);
-            stringCopy(status_line->text, status_label);
-            deleteString(status_label);
-        }
-        // fflush(stdout);
-
-        // Swap graphics buffers
-        // df::GraphicsManager &graphics_manager =
-        //     df::GraphicsManager::getInstance();
-        // graphics_manager.swapBuffers();
 
         // Adjust sleep time for additional framerate accuracy
         loop_time = split(&game_clock);
@@ -149,6 +166,8 @@ int loop() {
             adjust_time = 0;
         }
     }
+
+    closeGame(0);
 
     return 0;
 }
